[HUDI-5070] Move flaky cleaner tests to separate class (#7251)
